Add defined_bit option
With this option we can parse files that include a defined-bit in their
nodes. If this bit is on the function definition has found from the
analyzer (e.g. function of product). Otherwise, it means that the
definition was not found (e.g. function from shared library).
The format for the nodes is the following:

`[static|public]:[0|1]:path:function_name

0: not found
1: found