Replace ratio method with simpler solution
We don't wan to use a heuristic such as Levenshtein ratio. There are
three cases for products that contain shared libraries linked with a
binary.

1. Is a product that is included in the dependencies. In such case, we
simple need to match it with the dependency.
2. Is a product that is included in the dependencies but with slightly
different name. For example, libc6-udeb instead of libc6. In such cases,
we simply need to remove the -udeb to match the two product names. In
more complex cases where one product may have a dependency to
product-foo-dev and dpkg detect product-foo which is included in the
dependencies of product-foo-dev, we must find other solution.
3. The product is not included in the dependencies at all and we must
tag it as UNDEFINED.