Add entries to self.dependencies_lookup
This map will be used to set products in URIs. The following rules are
applied on how to set the products in the URIs.

* If a function is included in any one of some alternative dependencies,
then the highest alternative will be used as the product.
* If a functions is included in a product that provides a virtual
package that belongs to dependencies then the name of the virtual
product will be used as the product of the URI.